requirements.txt
langgraph
langchain
langchain-core
langchain-mistralai
langchain-openrouter
fastapi
uvicorn
pydantic
pydantic[email]
email-validator
pydantic-settings
sqlalchemy
psycopg2-binary
python-multipart
PyJWT
passlib[bcrypt]
pwdlib[argon2]
python-dotenv
pytest
httpx
mlflow
